ElectronicCatalog/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
drf-yasg
beautifulsoup4
requests
drf-spectacular
djangorestframework-simplejwt
django-filter
